VERSION 2 OF EXTPMF
ExtPMF has been completely refactored in order to use the most
recent versions of navtable, navtableforms and dbconnection, with
a new approach onto db instead of shp files.
We have also cleaned many classes and remade several operations
to work onto the db session instead of the TOC layers.
The forms new versions are still to be uploaded, but most of the
older ones still work.